servers/django-strawberry-async/pyproject.toml
[tool.poetry]
name = "django-graphene"
version = "0.1.0"
description = ""
authors = ["Jonathan Kim <jkimbo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"
strawberry-graphql = { git = "https://github.com/strawberry-graphql/strawberry.git", rev = "a45486e153d6db90650143bae8d521d7f6730f20" }
uvicorn = "^0.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
